Add an example for naming convetion for tests
Because of the dynamic parametrization, if the test is only identifiable with an id
such as "SOFTMAX/softmax_test.DISABLED_TestAll/59" for ex. it's hard to group test
failiures to identify patterns etc.

By replacing it with a naming scheme such as
"SOFTMAX/softmax_test.DISABLED_TestAll/59_DTf16_InputFMTother_InputDims400x600x2x2"
for ex. it's relatively obvious what the setup of the test was.

There's still a temp issue with scale however since the bias_term param isn't saved
during the tests setup. So that there's a need to refractor it somewhat and figure-
out how this should be done. This is somewhat similar to the issue with the static
init of the TEST_P params required and the uncomfortable setting of having
TearDownTestCase release the static setup values.